added optional fields, need to fix some jQuery hijinx.
doing a hide or show with no params works fine, but if you add a
duration, the div populates in the right column of the container and
slides under the existing fields on the form.  probably a pretty easy
fix.  also, the checkbox doesn't check.  this is also probably an easy
fix.